fix: ce144/145 validation and compact unit helper
